chg: dev: Start refactor: new dev manual !wip
I've starting diving hard into the refactor of the code base, ripping
out old code (it's always here in the repo, I can find it easy
enough.)

I'm deciding what to rewrite by starting with the "Manual of qtMUD
Development," and outlining the core concepts.

I'm fairly sure I know how I want to structure the core of the MUD,
and it's the spirit of what the "old" code is - I just now see that
the old code is poorly implemented.  (Well, considering how rapidly it
came together, I think it's pretty well done, but that's neither here
nor there.)